fix(web): route RPC settings test through playground proxy
